backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-adk>=1.20.0
google-genai>=1.0.0
google-cloud-firestore>=2.18.0
python-dotenv>=1.0.1
